checkpoint, added generic mapping functions

+++ src/kernel/include/mailbox/mailbox.h
@@ -214,7 +214,7 @@ uint32_t mailbox_clock_rate(clock_type ct);
 bool mailbox_generic_command(uint32_t tag_id, uint32_t id, uint32_t *value);
 bool mailbox_power_check(uint32_t type);
 bool mailbox_process(mailbox_tag *tag, uint32_t tag_size);
-void mailbox_write(uint8_t channel, uint32_t data);
+void mailbox_write(uint8_t channel, uint64_t data);
 uint32_t mailbox_read(uint8_t channel);
 
 #endif
